fix: address PR review feedback - thread safety, null guards, type fixes
- Add apiKey to in-flight request key in AvoEventSpecFetcher
- Remove unused java.util.concurrent import from AvoEventSpecFetchTypes
- Change double to long/int for timestamp/eventCount in EventSpecCacheEntry
- Add "unknown" streamId guard in AvoInspector constructor
- Defensive copy of eventProperties before async callback
- Add null guard for specResponse.metadata in handleBranchChangeAndCache
- Add null check for validationResult in bodyForValidatedEventSchemaCall
- Add timeouts and response logging to reportValidatedEvent
- Add size cap on EventSpecCache.set() to prevent unbounded growth
- Use ConcurrentHashMap for validator regex/allowedValues caches
- Wrap Double.parseDouble in try-catch in checkMinMaxRanges
- Fix tests: stub fetch callback with doAnswer for cache-miss paths
- Fix double casts in cache TTL tests

+++ avoinspector/src/main/java/app/avo/inspector/AvoEventSpecFetchTypes.java
@@ -1,7 +1,6 @@
 package app.avo.inspector;
 
 import java.util.*;
-import java.util.concurrent.*;
 
 class PropertyConstraintsWire {
     public String t;
@@ -58,9 +57,9 @@ class EventSpecResponse {
 
 class EventSpecCacheEntry {
     public EventSpecResponse spec;
-    public double timestamp;
-    public double lastAccessed;
-    public double eventCount;
+    public long timestamp;
+    public long lastAccessed;
+    public int eventCount;
 }
 
 class FetchEventSpecParams {
